BFRES: Shifted SMPJ dual-UV arrays to fix Ninji.
For whatever reason Ninji's model in SMPJ has conflicting "_u1" arrays which broke the script, this changes the dual-UV to use layer 3 (and the second pair to use 4 and 5) to prevent potential conflicts with other models. At some point I should rewrite this to ignore numbers and just write however many it finds in whatever order they happen to be in. And yes, this is a comically long description for a three-line change.